+++ src/main/java/com/example/paysvc/dto/Response/PaymentResponse.java
@@ -6,16 +6,18 @@ import jakarta.persistence.Enumerated;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PaymentResponse {
 
     private Long id;
-    private double amount;
-    private int userId;
-    private int merchantId;
-    private int debtId;
-    private int accountCode;
+    private BigDecimal amount;
+    private Integer userId;
+    private Integer merchantId;
+    private Integer debtId;
+    private Integer accountCode;
     private String idempotency;
     @Enumerated(EnumType.STRING)
     private Status status;
